refactor(cli): extract schema pack module
